requirements.txt
absl-py==2.0.0
addict==2.4.0
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
chumpy==0.70
click==8.1.3
cmake==3.25.0
comm==0.2.0
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==2.7.0
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.3
decorator==5.1.1
Deprecated==1.2.14
diffusers==0.22.3
einops==0.7.0
entrypoints==0.4
executing==2.0.1
fastjsonschema==2.18.1
ffmpeg-python==0.2.0
ffprobe==0.5
filelock==3.9.0
Flask==3.0.0
fonttools==4.44.0
fsspec==2023.10.0
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.23.4
google-auth-oauthlib==1.0.0
grpcio==1.59.2
h5py==3.10.0
huggingface-hub==0.19.0
hydra-core==0.11.3
idna==3.4
imageio==2.34.1
importlib-metadata==6.8.0
importlib-resources==6.1.1
iopath==0.1.10
ipycanvas==0.13.1
ipyevents==2.0.2
ipython==8.12.3
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
kornia==0.7.1
lit==15.0.7
lmdb==1.4.1
Markdown==3.5.1
MarkupSafe==2.1.1
marshmallow==3.20.2
matplotlib==3.7.3
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.7
msgpack-numpy==0.4.8
mypy-extensions==1.0.0
natsort==8.4.0
nbformat==5.7.0
nest-asyncio==1.5.9
networkx==3.0
ninja==1.11.1.1
numpy==1.23.1
nvisii==1.1.72
oauthlib==3.2.2
omegaconf==1.4.1
open3d==0.17.0
opencv-python==4.8.1.78
packaging==23.2
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
plotly==5.18.0
plyfile==1.0.1
portalocker==2.8.2
prompt-toolkit==3.0.39
protobuf==4.25.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pybullet==3.2.5
pyglet==1.5.28
pygltflib==1.16.1
Pygments==2.16.1
pyparsing==3.1.1
pyquaternion==0.9.9
python-dateutil==2.8.2
pytorch-lightning==0.7.1
pytz==2022.6
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
rpds-py==0.12.0
rsa==4.9
Rtree==1.1.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sk-video==1.1.10
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.3.0
threadpoolctl==3.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.13.0
transforms3d==0.4.1
trimesh==4.0.3
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.4.0
urllib3==1.26.13
usd-core==23.5
wcwidth==0.2.9
web-annotation-tool==0.0.1
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
yacs==0.1.8
zipp==3.10.0
